requirements.txt
accelerate==0.32.1
agentql==1.6.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.6.2.post1
APScheduler==3.11.0
async-timeout==5.0.1
attrs==24.2.0
av==12.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.35.60
botocore==1.35.66
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
dateparser==1.2.0
defusedxml==0.7.1
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
einops==0.8.0
fake-http-header==0.3.5
fastapi==0.110.3
fastembed==0.4.2
filelock==3.16.1
filetype==1.2.0
firebase-admin==6.5.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.55.0
free_proxy==1.1.3
frozenlist==1.5.0
fsspec==2024.10.0
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
googlesearch-python==1.2.5
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.68.0
grpcio-status==1.67.1
grpcio-tools==1.68.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ifaddr==0.2.0
imageio==2.36.0
inference-gpu==0.28.2
inference-sdk==0.28.2
iniconfig==2.0.0
instaloader==4.14
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.1
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.7
langchain-aws==0.2.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-google-genai==2.0.5
langchain-mistralai==0.2.2
langchain-ollama==0.2.0
langchain-openai==0.2.9
langchain-text-splitters==0.3.2
langsmith==0.1.144
lazy_loader==0.4
loguru==0.7.2
lxml==5.3.0
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.23.1
matplotlib==3.9.2
minify_html==0.15.0
mistral_common==1.5.1
mmh3==4.1.0
mockito==1.5.3
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
ollama==0.4.0
onnx==1.17.0
onnxruntime==1.19.2
onnxruntime-gpu==1.19.2
openai==1.54.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.11
packaging==24.1
pandas==2.2.3
peft==0.11.1
piexif==1.1.3
pillow==10.4.0
playwright==1.48.0
pluggy==1.5.0
portalocker==2.10.1
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.0.2
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydot==2.0.0
pyee==12.0.0
Pygments==2.18.0
PyJWT==2.9.0
pylibsrtp==0.10.0
pyOpenSSL==24.2.1
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mockito==0.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.1
redis==5.0.8
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.0.1
roboflow==1.1.49
rpds-py==0.21.0
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-image==0.24.0
scipy==1.14.1
scrapegraphai==1.31.0
semchunk==2.2.0
sentencepiece==0.2.0
shapely==2.0.6
shellingham==1.5.4
simpleeval==1.0.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
starlette==0.37.2
structlog==24.4.0
supervision==0.22.0
sympy==1.13.3
tenacity==9.0.0
tf-playwright-stealth==1.1.0
tifffile==2024.9.20
tiktoken==0.7.0
timm==1.0.11
tldextract==5.1.3
tokenizers==0.20.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.67.0
transformers==4.46.3
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
undetected-playwright==0.3.0
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.18.0
zxing-cpp==2.2.0
